+++ tests/requirements/compile.py
@@ -7,6 +7,60 @@ import sys
 from pathlib import Path
 
 
+COMPILE_SETTINGS = {
+    "py39-dj32-cms41": [],
+    "py310-dj32-cms41": [],
+    "py39-dj40-cms41": [],
+    "py310-dj40-cms41": [],
+    "py311-dj40-cms41": [],
+    "py39-dj41-cms41": [],
+    "py310-dj41-cms41": [],
+    "py311-dj41-cms41": [],
+    "py39-dj42-cms41": [],
+    "py310-dj42-cms41": [],
+    "py311-dj42-cms41": [],
+    "py311-djmain-cms41": [],
+}
+
+django_dict = {
+    "djmain": "https://github.com/django/django/tarball/main#egg=Django",
+    "dj32": "Django>=3.2,<4",
+    "dj40": "Django>=4.0,<4.1",
+    "dj41": "Django>=4.1,<4.2",
+    "dj42": "Django>=4.2,<5",
+}
+
+cms_dict = {
+    "cms40": "https://github.com/django-cms/django-cms/tarball/release/4.0.1.x#egg=django-cms",
+    "cms41": "django-cms>=4.1.0rc2,<4.2",
+}
+
+
+def get_args(key, value, common_args):
+    py_ver, dj_ver, cms_ver, mode = key.split("-")
+    assert py_ver[:2] == "py"
+    assert mode.endswith(".txt")
+    return [
+        f"python{py_ver[2]}.{py_ver[3:]}",
+        *common_args,
+        "-P",
+        django_dict[dj_ver],
+        "-P",
+        cms_dict[cms_ver],
+        *value,
+        "-o",
+        key
+    ]
+
+
+def run(*args, **kwargs):
+    try:
+        print(" ".join(args[0]))
+        subprocess.run(*args, **kwargs)
+    except Exception as e:
+        print(f'Failed {" ".join(args[0])}: {e}')
+
+
 if __name__ == "__main__":
     os.chdir(Path(__file__).parent)
     os.environ["CUSTOM_COMPILE_COMMAND"] = "requirements/compile.py"
@@ -15,206 +69,18 @@ if __name__ == "__main__":
         "-m",
         "piptools",
         "compile",
-        # "--generate-hashes",
+        # temporarily remove "--generate-hashes", until all dependencies are actual releases
         "--allow-unsafe",
     ] + sys.argv[1:]
-    subprocess.run(
-        [
-            "python3.8",
-            *common_args,
-            "-P",
-            "Django>=3.2,<3.3",
-            "-P",
-            "django-treebeard==4.5.1",
-            "-P",
-            "django-classy-tags<4.0",
-            "-o",
-            "py38-dj32-cms40-default.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.8",
-            *common_args,
-            "-P",
-            "Django>=3.2,<3.3",
-            "-P",
-            "django-treebeard==4.5.1",
-            "-P",
-            "django-classy-tags<4.0",
-            "-o",
-            "py38-dj32-cms40-versioning.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.8",
-            *common_args,
-            "-P",
-            "Django>=3.2,<3.3",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py38-dj32-cms41-default.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.8",
-            *common_args,
-            "-P",
-            "Django>=3.2,<3.3",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py38-dj32-cms41-versioning.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.9",
-            *common_args,
-            "-P",
-            "Django>=3.2,<3.3",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py39-dj32-cms41-default.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.9",
-            *common_args,
-            "-P",
-            "Django>=3.2,<3.3",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py39-dj32-cms41-versioning.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.9",
-            *common_args,
-            "-P",
-            "Django>=4.0,<4.1",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py39-dj40-cms41-default.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.9",
-            *common_args,
-            "-P",
-            "Django>=4.0,<4.1",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py39-dj40-cms41-versioning.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.10",
-            *common_args,
-            "-P",
-            "Django>=3.2,<3.3",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py310-dj32-cms41-default.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.10",
-            *common_args,
-            "-P",
-            "Django>=3.2,<3.3",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py310-dj32-cms41-versioning.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.10",
-            *common_args,
-            "-P",
-            "Django>=4.0,<4.1",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py310-dj40-cms41-default.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.10",
-            *common_args,
-            "-P",
-            "Django>=4.0,<4.1",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py310-dj40-cms41-versioning.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.11",
-            *common_args,
-            "-P",
-            "Django>=4.1,<4.2",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py311-dj41-cms41-default.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    subprocess.run(
-        [
-            "python3.11",
-            *common_args,
-            "-P",
-            "Django>=4.1,<4.2",
-            "-P",
-            "django-cms==4.1rc1",
-            "-o",
-            "py311-dj41-cms41-versioning.txt",
-        ],
-        check=True,
-        capture_output=True,
-    )
+
+    for key, value in COMPILE_SETTINGS.items():
+        run(
+            get_args(key + "-default.txt", value, common_args),
+            check=True,
+            capture_output=True,
+        )
+        run(
+            get_args(key + "-versioning.txt", value, common_args),
+            check=True,
+            capture_output=True,
+        )