+++ tests/test_admin.py
@@ -12,10 +12,15 @@ from cms.utils.urlutils import add_url_parameters, admin_reverse
 from bs4 import BeautifulSoup
 
 from djangocms_alias.constants import (
-    CHANGE_ALIASCONTENT_URL_NAME,
+    CHANGE_ALIAS_URL_NAME,
     USAGE_ALIAS_URL_NAME,
 )
-from djangocms_alias.models import Alias as AliasModel, AliasContent, Category
+from djangocms_alias.models import (
+    Alias,
+    Alias as AliasModel,
+    AliasContent,
+    Category,
+)
 from djangocms_alias.utils import is_versioning_enabled
 from tests.base import BaseAliasPluginTestCase
 
@@ -100,7 +105,7 @@ class AliasContentManagerTestCase(BaseAliasPluginTestCase):
         )
 
     @skipUnless(is_versioning_enabled(), 'Test only relevant for versioning')
-    def test_alias_content_manager_rendering_with_versioning_actions(self):
+    def test_alias_changelist_rendering_with_versioning_actions(self):
         """
         When rendering aliascontent manager with versioning actions
         """
@@ -115,14 +120,13 @@ class AliasContentManagerTestCase(BaseAliasPluginTestCase):
             language="en",
         )
 
-        from djangocms_versioning.helpers import proxy_model
         from djangocms_versioning.models import Version
 
-        alias_version = Version.objects.create(content=expected_en_content, created_by=self.superuser)
+        Version.objects.create(content=expected_en_content, created_by=self.superuser)
 
         with self.login_user_context(self.superuser):
 
-            base_url = self.get_admin_url(AliasContent, "changelist")
+            base_url = self.get_admin_url(Alias, "changelist")
             # en is the default language configured for the site
             response = self.client.get(base_url)
 
@@ -170,31 +174,17 @@ class AliasContentManagerTestCase(BaseAliasPluginTestCase):
             latest_alias_content_version.get_state_display(),
             response_content_decoded,
         )
-
         self.assertIn(
             localize(localtime(latest_alias_content_version.modified)),
             response_content_decoded,
         )
 
         usage_url = admin_reverse(USAGE_ALIAS_URL_NAME, args=[expected_en_content.alias.pk])
-        change_category_and_site_url = admin_reverse(
+        settings_url = admin_reverse(
             '{}_{}_change'.format(
                 expected_en_content._meta.app_label,
                 expected_en_content.alias._meta.model_name
             ), args=(expected_en_content.alias.pk,)
-        )
-        rename_alias_url = admin_reverse(
-            '{}_{}_change'.format(
-                expected_en_content._meta.app_label,
-                expected_en_content._meta.model_name
-            ), args=(expected_en_content.pk,)
-        )
-        version = proxy_model(alias_version, expected_en_content)
-        edit_alias_url = admin_reverse(
-            "{app}_{model}_edit_redirect".format(
-                app=version._meta.app_label, model=version._meta.model_name
-            ),
-            args=(version.pk,),
         )
 
         self.assertIn(
@@ -202,16 +192,8 @@ class AliasContentManagerTestCase(BaseAliasPluginTestCase):
             response_content_decoded,
         )
         self.assertIn(
-            rename_alias_url,
+            settings_url,
             response_content_decoded,
-        )
-        self.assertIn(
-            change_category_and_site_url,
-            response_content_decoded,
-        )
-        self.assertIn(
-            edit_alias_url,
-            response_content_decoded
         )
 
     @skipUnless(is_versioning_enabled(), 'Test only relevant for versioning')
@@ -230,30 +212,18 @@ class AliasContentManagerTestCase(BaseAliasPluginTestCase):
             language="en",
         )
 
-        from djangocms_versioning.helpers import proxy_model
         from djangocms_versioning.models import Version
 
-        alias_version = Version.objects.create(content=expected_en_content, created_by=self.superuser)
-        version = proxy_model(alias_version, expected_en_content)
-        edit_alias_url = admin_reverse(
-            "{app}_{model}_edit_redirect".format(
-                app=version._meta.app_label, model=version._meta.model_name
-            ),
-            args=(version.pk,),
-        )
+        Version.objects.create(content=expected_en_content, created_by=self.superuser)
 
         with self.login_user_context(self.superuser):
-            base_url = self.get_admin_url(AliasContent, "changelist")
+            base_url = self.get_admin_url(Alias, "changelist")
             # en is the default language configured for the site
             response = self.client.get(base_url)
         response_content_decoded = response.content.decode()
 
         self.assertIn(
             expected_en_content.get_absolute_url(),
-            response_content_decoded,
-        )
-        self.assertIn(
-            edit_alias_url,
             response_content_decoded,
         )
         self.assertNotIn(
@@ -314,13 +284,13 @@ class AliasContentManagerTestCase(BaseAliasPluginTestCase):
         Version.objects.create(content=last_alias_content, created_by=self.superuser)
 
         with self.login_user_context(self.superuser):
-            base_url = self.get_admin_url(AliasContent, "changelist")
+            base_url = self.get_admin_url(Alias, "changelist")
             # o=1 indicates ascending alphabetical order on list_displays second entry
             base_url += "?o=1"
             # en is the default language configured for the site
             response = self.client.get(base_url)
         soup = BeautifulSoup(response.content, "html.parser")
-        results = soup.find_all("td", class_="field-get_category")
+        results = soup.find_all("td", class_="field-category")
 
         # Test results are in ascending alphabetical order
         self.assertEqual(results[0].text, first_alias_content.name)
@@ -330,14 +300,13 @@ class AliasContentManagerTestCase(BaseAliasPluginTestCase):
         self.assertEqual(results[4].text, last_alias_content.name)
 
         with self.login_user_context(self.superuser):
-            base_url = self.get_admin_url(AliasContent, "changelist")
+            base_url = self.get_admin_url(Alias, "changelist")
             # o=-1 indicates descending alphabetical order on list_displays second entry
             base_url += "?o=-1"
             # en is the default language configured for the site
             response = self.client.get(base_url)
         soup = BeautifulSoup(response.content, "html.parser")
-        results = soup.find_all("td", class_="field-get_category")
-
+        results = soup.find_all("td", class_="field-category")
         # Test results are in descending alphabetical order
         self.assertEqual(results[4].text, first_alias_content.name)
         self.assertEqual(results[3].text, first_alias_content_lower.name)
@@ -437,7 +406,7 @@ class AliasContentManagerTestCase(BaseAliasPluginTestCase):
 
         with self.login_user_context(self.superuser):
             response = self.client.get(
-                self.get_list_aliascontent_endpoint(),
+                self.get_list_alias_endpoint(),
             )
 
         self.assertEqual(response.status_code, 200)
@@ -467,13 +436,13 @@ class AliasContentManagerTestCase(BaseAliasPluginTestCase):
             alias2_content = alias2.get_content(language=self.language)
             alias3_content = alias3.get_content(language=self.language)
             aliascontent1_url = admin_reverse(
-                CHANGE_ALIASCONTENT_URL_NAME, args=[alias1_content.pk]
+                CHANGE_ALIAS_URL_NAME, args=[alias1_content.pk]
             )
             aliascontent2_url = admin_reverse(
-                CHANGE_ALIASCONTENT_URL_NAME, args=[alias2_content.pk]
+                CHANGE_ALIAS_URL_NAME, args=[alias2_content.pk]
             )
             aliascontent3_url = admin_reverse(
-                CHANGE_ALIASCONTENT_URL_NAME, args=[alias3_content.pk]
+                CHANGE_ALIAS_URL_NAME, args=[alias3_content.pk]
             )
 
         self.assertContains(response, aliascontent1_url)
@@ -591,8 +560,8 @@ class AliasesManagerTestCase(BaseAliasPluginTestCase):
 
         response = self.client.get(index_url)
 
-        unexpected_content = '<th scope="row"><a href="/en/admin/djangocms_alias/alias/">Aliases</a></th>'
-        expected_content = '<th scope="row"><a href="/en/admin/djangocms_alias/aliascontent/">Alias contents</a></th>'
+        unexpected_content = '<th scope="row"><a href="/en/admin/djangocms_alias/aliascontent/">Alias contents</a></th>'
+        expected_content = '<th scope="row"><a href="/en/admin/djangocms_alias/alias/">Aliases</a></th>'
 
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, expected_content)